feat: 2단계 - 요구사항2 구현
