+++ nextstep-mvc/src/main/java/nextstep/mvc/Controller.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.asis;
+package nextstep.mvc;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
