+++ nextstep-mvc/src/main/java/nextstep/mvc/ControllerScanner.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc;
 
 import org.reflections.Reflections;
