+++ nextstep-mvc/src/main/java/nextstep/mvc/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc;
 
 import nextstep.web.annotation.RequestMethod;
