--- nextstep-mvc/src/main/java/nextstep/mvc/asis/ForwardController.java
@@ -1,20 +1,0 @@
-package nextstep.mvc.asis;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class ForwardController implements Controller {
-    private String forwardUrl;
-
-    public ForwardController(String forwardUrl) {
-        this.forwardUrl = forwardUrl;
-        if (forwardUrl == null) {
-            throw new NullPointerException("forwardUrl is null. 이동할 URL을 입력하세요.");
-        }
-    }
-
-    @Override
-    public String handle(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        return forwardUrl;
-    }
-}
