+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/NotSupportedViewTypeException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.exception;
+package nextstep.mvc.exception;
 
 public class NotSupportedViewTypeException extends RuntimeException {
     public NotSupportedViewTypeException(String message) {
