+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/ObjectMapperException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe.exception;
+package nextstep.mvc.exception;
 
 public class ObjectMapperException extends RuntimeException {
     public ObjectMapperException() {
