+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/JsonViewTest.java
@@ -1,7 +1,7 @@
 package nextstep.mvc.tobe;
 
-import nextstep.mvc.tobe.view.JsonView;
-import nextstep.mvc.tobe.view.View;
+import nextstep.mvc.view.JsonView;
+import nextstep.mvc.view.View;
 import nextstep.utils.JsonUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
