--- slipp/src/main/java/slipp/controller/UpdateFormUserController.java
@@ -1,22 +1,0 @@
-package slipp.controller;
-
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-import nextstep.mvc.asis.Controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class UpdateFormUserController implements Controller {
-
-    @Override
-    public String handle(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        User user = DataBase.findUserById(userId);
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
-        }
-        req.setAttribute("user", user);
-        return "/user/updateForm.jsp";
-    }
-}
