+++ slipp/src/main/java/slipp/controller/tobe/AuthenticateController.java
@@ -1,8 +1,8 @@
 package slipp.controller.tobe;
 
-import nextstep.mvc.tobe.view.JspView;
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.mvc.tobe.view.RedirectView;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.mvc.view.RedirectView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
@@ -15,7 +15,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 @Controller
-public class LoginController {
+public class AuthenticateController {
     @RequestMapping(value = "/users/login", method = RequestMethod.POST)
     public ModelAndView login(HttpServletRequest req, HttpServletResponse resp) {
         String userId = req.getParameter("userId");
@@ -33,5 +33,12 @@ public class LoginController {
             req.setAttribute("loginFailed", true);
             return new ModelAndView(new JspView("/user/login.jsp"));
         }
+    }
+
+    @RequestMapping(value = "/users/logout", method = RequestMethod.GET)
+    public ModelAndView logout(HttpServletRequest req, HttpServletResponse resp) {
+        HttpSession session = req.getSession();
+        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
+        return new ModelAndView(new RedirectView("/"));
     }
 }
