--- slipp/src/main/java/slipp/controller/tobe/UserController.java
@@ -1,53 +1,0 @@
-package slipp.controller.tobe;
-
-import nextstep.mvc.tobe.view.JspView;
-import nextstep.mvc.tobe.view.ModelAndView;
-import nextstep.mvc.tobe.view.RedirectView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import slipp.controller.UserSessionUtils;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class UserController {
-    private static final Logger log = LoggerFactory.getLogger(UserController.class);
-
-    @RequestMapping(value = "/users", method = RequestMethod.GET)
-    public ModelAndView userList(HttpServletRequest req, HttpServletResponse resp) {
-        if (!UserSessionUtils.isLogined(req.getSession())) {
-            return new ModelAndView(new RedirectView("/users/loginForm"));
-        }
-        req.setAttribute("users", DataBase.findAll());
-        return new ModelAndView(new JspView("/user/list.jsp"));
-    }
-
-    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
-    public ModelAndView userForm(HttpServletRequest req, HttpServletResponse resp) {
-        return new ModelAndView(new JspView("/user/form.jsp"));
-    }
-
-    @RequestMapping(value = "/users/update", method = RequestMethod.PUT)
-    public ModelAndView userUpdate(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        User user = DataBase.findUserById(req.getParameter("userId"));
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
-        }
-
-        User updateUser = new User(
-                req.getParameter("userId"),
-                req.getParameter("password"),
-                req.getParameter("name"),
-                req.getParameter("email")
-        );
-        log.debug("Update User : {}", updateUser);
-        user.update(updateUser);
-        return new ModelAndView(new RedirectView("/"));
-    }
-}
