+++ slipp/src/main/java/slipp/controller/tobe/UserFormController.java
@@ -1,0 +1,41 @@
+package slipp.controller.tobe;
+
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import slipp.controller.UserSessionUtils;
+import slipp.domain.User;
+import slipp.support.db.DataBase;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@Controller
+public class UserFormController {
+    private static final Logger log = LoggerFactory.getLogger(UserFormController.class);
+
+    @RequestMapping(value = "/users/form", method = RequestMethod.GET)
+    public ModelAndView userForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView(new JspView("/user/form.jsp"));
+    }
+
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public ModelAndView loginForm(HttpServletRequest req, HttpServletResponse resp) {
+        return new ModelAndView(new JspView("/user/login.jsp"));
+    }
+
+    @RequestMapping(value = "/users/updateForm", method = RequestMethod.GET)
+    public ModelAndView updateForm(HttpServletRequest req, HttpServletResponse resp) {
+        String userId = req.getParameter("userId");
+        User user = DataBase.findUserById(userId);
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+        req.setAttribute("user", user);
+        return new ModelAndView(new JspView("/user/updateForm.jsp"));
+    }
+}
