+++ slipp/src/test/java/slipp/controller/tobe/UserFormControllerTest.java
@@ -1,12 +1,11 @@
 package slipp.controller.tobe;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpMethod;
 import support.test.NsWebTestClient;
 
-class UserControllerTest {
+class UserFormControllerTest {
     private NsWebTestClient client;
 
     @BeforeEach
